commit atuaizado(certo)

+++ praticas/praticas1/questao2.c
@@ -3,17 +3,17 @@ Faça um programa em C que calcule a área de um triângulo (a = base x altura /
 */
 #include  <stdio.h>
 int main(){
-  int base;
-  int altura;
+  float base;
+  float altura;
 
   printf("Digite a base do triângulo: ");
-  int deu_certo = scanf("%i", &base);
+  int deu_certo = scanf("%f", &base);
 
   printf("Digite a altura do triângulo: ");
-  int deu_certo2 = scanf("%i", &altura);
+  int deu_certo2 = scanf("%f", &altura);
 
-  int area = (base * altura) / 2;
-  printf("A área do triângulo é: %i\n", area);
+  float area = (base * altura) / 2;
+  printf("A área do triângulo é: %f\n", area);
   
   return   0;
 }
